PaaS_Clever_cloud/Backend/Poster/requirements.txt
fastapi
motor
python-dotenv
uvicorn
pydantic
python-multipart
aiohttp
google-cloud-storage
google-cloud-pubsub
dnspython